Backdoor commands #14
A recent issue I encountered as it related to YAHOO returning a stock price that is very old.  The way 3'rd party sites are used to find prices depends on the speed (and valid data) - so that a specific site is tried first.  So if YAHOO was set as the favorite to find "TSE:CJP" - it was messing up my portfolio values.  I added some commands that allow me to set the default primary site and also the blocked site.  So for example, you can query which site is currently used and which is permanently blocked like:   =CACHEFINANCE("TSE:CJP", "Price", "GETBLOCKED")  or which is the primary lookup site :  =CACHEFINANCE("TSE:CJP", "Price", "GET") .  You can then set the permanent site to block like this:  =CACHEFINANCE("TSE:CJP", "Price", "SETBLOCKED", "YAHOO")

to list all valid commands:  =CACHEFINANCE("", "", "?")

Which lists the commands as:

    ? (display help)
    TEST (tests web sites)
    CLEARCACHE (remove cache - run again if timeout. If symbol/attribute blank - removes all)
    REMOVE (pref. site set as do not use site for symbol/attribute)
    LIST (show all supported web lookups)
    GET (current pref. site for symbol/attribute)
    GETBLOCKED (current blocked site for symbol/attribute)
    SET (4'th parm is set to pref. site for symbol/attribute)
    SETBLOCKED (4'th parm is set to blocked site for symbol/attribute)