Cache Trigger removed.
The purpose of the trigger was to handle large numbers of stock  because the max runtime for triggers was 5 minutes as opposed to 30 seconds for a custom function.
I have modified how I approach a large volume of stocks in the custom function and as a result - the custom function if far and away more reliable than trying to use the very inconsistent google sheets trigger functions.  If you want access to trigger code, view the history of this project.