#15.  Empty DEFAULT parameter --> No cache lookup.
The empty default parameter was affecting the "yieldpct" attribute (googefinance does not have this value for ETF's) - but the CACHEFINANCES() was using a map on that array to fill the values for all the yields.  However, this array empty - so nothing was copied from the short cache.  So if a future trigger ran the "yieldpct" lookup before the expected cache timeout - it would do the actual web lookups for all the tickers.