#16.  Single ticker for CACHEFINANCES().
If currency lookup fails as default input, it will use AlphaVantage to extract exchange rate.  NOTE:  You must have an AlphaVantage key and it must be set in the project properties as ALPHA_VANTAGE_API_KEY.  Also note, this FREE api is limited to 25 requests per day, so don't expect it to work if requesting dozens of different exchange rates if googlefinance() is failing.