Added TwelveData rest API.  #18 Basic throttling added
Add to script settings 'TWELVE_DATA_API_KEY' your key from https://twelvedata.com/
Basic SIMPLE throttling for sites that have use limits.  So for example, if site has a limit of 8 calls per second, the function tracks the number of uses in second as returned by the system clock.  So this approach will fail IF for example 5 calls in the second of '22' and 5 calls in the second of '23' - it would NOT throttle because it did not exceed in either '22' or '23' - but the site may have a rolling limit - doing a rolling limit is more complicated - so not used here.  
Added a new method which is required for any new site added called getThrottleObject()