Duplicate processing for duplicate symbol(s) on 3'rd party lookup.
In the case where the same stock symbol appeared more than once in a CACHEFINANCES() lookup of many stocks and a 3'rd party web lookup was required - it was submitting the second (or more) symbol again for lookups.  Where this was an issue is when multiple sites (searched in order) all failed and it was a site later in the processing order that actually found that data - this whole process was repeated again for the failing sites.  The process was working - it was just not optimal.
ALSO:  added the YahooApi lookup into the list of sites doing screen scraping (in this case the api returns JSON - which is much faster).